React project+base app configured and working
this is a barebones base of what we plan to work on. using 'npm start' with the project as cd on CMD should put a local server with the base app working in a firefox browser (Edge for some reason isn't able to read the script properly, will have to look that later). that being said right now the app isn't really using React for anything besides the server