folder creation method created and added to the start
the part that created the folder if it didn't exist was in an akward spot, now it should start every time a chat is started. also the basechat and chaturl variables are now global to the service. on the other hand, now we have to go back to the publick folder because of permission denied errors